Rename "input_keyboard_command.dart" to "input_command.dart".
