+++ patterns/command/commands/input_command.dart
